Fix rounding issue

+++ pyflam3ng/func.py
@@ -39,5 +39,5 @@ def hls2rgb(h, l, s):
     h = clip(h, 0, 1, True)
     l = clip(l, 0, 1)
     s = clip(s, 0, 1)
-    return map(lambda x: int(x*256), colorsys.hls_to_rgb(h, l, s))
+    return map(lambda x: x*256, colorsys.hls_to_rgb(h, l, s))
